requirements.txt
Django==3.1
django-environ==0.4.5
gunicorn==20.0.4
psycopg2==2.8.5
whitenoise==5.2.0
djangorestframework==3.12.4
Markdown==3.3.4
django-filter==21.1
